requirements.txt
TensorFlowTTS
gradio 
numpy
SoundFile
git+https://github.com/repodiac/german_transliterate
tensorflow-cpu